simple refact.

+++ sourceCode/platform/vArchive.cpp
@@ -20,7 +20,7 @@ VArchive::~VArchive()
 
 void VArchive::writeString(const std::string& str)
 {
-    for (int i=0;i<int(str.length());i++)
+    for (size_t i = 0; i < str.length(); i++)
         (*this) << str[i];
 }
